art_guessing/requirements.txt
matplotlib
numpy
requests
fastapi
uvicorn
python-multipart
opencv-python

#image processing
pillow

#Data processing and models
tensorflow
